Refactor gsheet refresh endpoints & adjust gsheet utils calls
