+++ src/gsheet/refresh/router.py
@@ -8,34 +8,17 @@ import gspread
 from src.config import settings
 from src.database.postgres.core import make_session
 from src.database.postgres.core import engine as CONN
-import src.gsheet.refresh.service as service
 from src.gsheet.refresh.attendance.router import router as attendance_router
+from src.gsheet.refresh.main.router import router as main_router
 
 router = APIRouter()
 
-@router.post("/main",
-    description="Refresh main roster information on the associated Google Sheet",
-    response_description="Updated main roster",
-    status_code=status.HTTP_201_CREATED)
-def refresh_main(db: Session = Depends(make_session)) -> Dict[str, Any]:
-    """
-    Copy database records to the specified Google Sheet (Main Roster)
-    """
-    try:
-        # Pull roster & write to sheet
-        roster_data = service.fetch_roster(CONN)
-        if settings.app_env == "production":
-            gc = service.create_credentials()
-            return service.write_to_gsheet(roster_data, "Main Roster", gc, environ.get("ROSTER_SHEET_KEY"))
-        else:
-            gc = gspread.service_account(filename='gspread_credentials.json')
-            return service.write_to_gsheet(roster_data, "Main Roster", gc, environ.get("ROSTER_SHEET_KEY"))
-
-    except SQLAlchemyError as e:
-        db.rollback()
-        raise HTTPException(status_code=500, detail=str(e))
+router.include_router(
+    main_router,
+    prefix="/main"
+)
 
 router.include_router(
     attendance_router,
-    prefix="/attendance",
+    prefix="/attendance"
 )
